refactor: extract tweet composer logic to a composable
