+++ app/components/tweet/composer/TweetComposer.vue
@@ -1,15 +1,13 @@
 <script setup lang="ts">
-import { ref, computed, defineEmits } from 'vue';
+import { ref, defineEmits, toRef } from 'vue';
 import { useUserStore } from '@/stores/user';
 import TweetEditor from './TweetEditor.vue';
 import Toolbar from './Toolbar.vue';
 import MediaSlideshow from './MediaSlideshow.vue';
 import type { MediaItem } from '~~/shared/types/shared';
 import type { Tweet } from '~~/shared/types/tweets';
-import { uploadMediaService } from '@/services/tweet/uploadMediaService';
-import { createTweetService } from '@/services/tweet/createTweetService';
-import { showToaster } from '@/utils/showToaster';
 import Avatar from '~/components/ui/Avatar.vue';
+import { usePostTweet } from '@/composables/usePostTweet';
 
 interface Props {
   replyToTweetId?: string | null;
@@ -26,115 +24,34 @@ const tweetEditorRef = ref<InstanceType<typeof TweetEditor> | null>(null);
 const userStore = useUserStore();
 const media = ref<MediaItem[]>([]);
 
-// Loading states
-const isPosting = ref(false);
-const uploadProgress = ref(0);
-const uploadStatus = ref<'uploading' | 'posting' | null>(null);
-
-const MAX_LENGTH = 280;
-const MAX_MEDIA = 4;
-
-const characterCount = computed(() => tweetContent.value.length);
-const isOverLimit = computed(() => characterCount.value > MAX_LENGTH);
-
-const loadingMessage = computed(() => {
-  if (uploadStatus.value === 'uploading') {
-    return $t('tweet.composer.loading.uploading', { progress: uploadProgress.value });
-  }
-  if (uploadStatus.value === 'posting') {
-    return $t('tweet.composer.loading.posting');
-  }
-  return '';
-});
-
-const { uploadImage, uploadVideo } = uploadMediaService();
+const replyToRef = toRef(props, 'replyToTweetId');
+const {
+  isPosting,
+  MAX_LENGTH,
+  MAX_MEDIA,
+  characterCount,
+  isOverLimit,
+  loadingMessage,
+  handlePost,
+  handleAddMedia,
+  handleRemoveMedia,
+} = usePostTweet(tweetContent, media, replyToRef);
 
 const emit = defineEmits<{
   (e: 'posted', tweet: Tweet): void;
 }>();
 
-const handlePost = async () => {
-  if (!tweetContent.value.trim() && media.value.length === 0) return;
-  if (isOverLimit.value) return;
-  if (isPosting.value) return; // Prevent double submission
+const handlePostWrapper = async () => {
+  const newTweet = await handlePost();
+  if (!newTweet) return;
 
-  isPosting.value = true;
-  uploadProgress.value = 0;
+  emit('posted', newTweet);
 
-  try {
-    // Upload media files → get media IDs
-    const mediaIds: string[] = [];
-
-    if (media.value.length > 0) {
-      uploadStatus.value = 'uploading';
-
-      for (let i = 0; i < media.value.length; i++) {
-        const item = media.value[i];
-
-        if (item?.type === 'image') {
-          const mediaId = await uploadImage(item.file, 'tweets');
-          mediaIds.push(mediaId);
-        } else if (item?.type === 'video') {
-          const mediaId = await uploadVideo(item?.file, 'tweets');
-          mediaIds.push(mediaId);
-        }
-
-        // Update progress
-        uploadProgress.value = Math.round(((i + 1) / media.value.length) * 100);
-      }
-    }
-
-    // Send create tweet request
-    uploadStatus.value = 'posting';
-    const newTweet = await createTweetService({
-      content: tweetContent.value,
-      media: mediaIds,
-      replyToTweetId: props.replyToTweetId,
-    });
-
-    // eslint-disable-next-line no-console
-    console.log('Tweet created:', newTweet);
-
-    showToaster('success', 'Tweet posted successfully!');
-
-    // Emit the created tweet so parent can handle it appropriately
-    emit('posted', newTweet);
-
-    // Cleanup
-    media.value.forEach((item) => URL.revokeObjectURL(item.url));
-    tweetContent.value = '';
-    media.value = [];
-    tweetEditorRef.value?.resetHeight();
-  } catch {
-    showToaster('error', 'Error creating tweet. Please try again.');
-  } finally {
-    isPosting.value = false;
-    uploadStatus.value = null;
-    uploadProgress.value = 0;
-  }
-};
-
-const handleAddMedia = (files: File[]) => {
-  files.forEach((file) => {
-    if (media.value.length === MAX_MEDIA) return;
-
-    const url = URL.createObjectURL(file);
-    const id = `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-
-    const type = file.type.startsWith('video') ? 'video' : 'image';
-
-    media.value.push({ id, file, url, type });
-  });
-};
-
-const handleRemoveMedia = (id: string) => {
-  const index = media.value.findIndex((m) => m.id === id);
-
-  if (index !== -1 && index < media.value.length && media.value[index]) {
-    // Revoke the blob URL to free memory
-    URL.revokeObjectURL(media.value[index].url);
-    media.value.splice(index, 1);
-  }
+  // Cleanup
+  media.value.forEach((item) => URL.revokeObjectURL(item.url));
+  tweetContent.value = '';
+  media.value = [];
+  tweetEditorRef.value?.resetHeight();
 };
 </script>
 
@@ -177,7 +94,7 @@ const handleRemoveMedia = (id: string) => {
       :can-add-media="media.length < MAX_MEDIA"
       :button-text="$t('tweet.composer.button.' + props.type)"
       :is-posting="isPosting"
-      @post="handlePost"
+      @post="handlePostWrapper"
       @add-media="handleAddMedia"
     />
   </div>
